+++ src/lib/pages/instantiate/instantiate.tsx
@@ -1,5 +1,6 @@
-import { Heading, Text } from "@chakra-ui/react";
+import { Flex, Heading, Text } from "@chakra-ui/react";
 import type { InstantiateResult } from "@cosmjs/cosmwasm-stargate";
+import type { StdFee } from "@cosmjs/stargate";
 import { useWallet } from "@cosmos-kit/react";
 import { useQuery } from "@tanstack/react-query";
 import Long from "long";
@@ -9,20 +10,22 @@ import { useForm } from "react-hook-form";
 
 import {
   useFabricateFee,
-  useSimulateFee,
   useInstantiateTx,
   useCelatoneApp,
   useLCDEndpoint,
   useValidateAddress,
+  useSimulateFeeQuery,
 } from "lib/app-provider";
 import { AssignMe } from "lib/components/AssignMe";
 import { ConnectWalletAlert } from "lib/components/ConnectWalletAlert";
+import { EstimatedFeeRender } from "lib/components/EstimatedFeeRender";
 import type { FormStatus } from "lib/components/forms";
 import { ControllerInput } from "lib/components/forms";
 import { AttachFund } from "lib/components/fund";
 import { defaultAsset, defaultAssetJsonStr } from "lib/components/fund/data";
 import type { AttachFundsState } from "lib/components/fund/types";
 import { AttachFundsType } from "lib/components/fund/types";
+import { CustomIcon } from "lib/components/icon";
 import JsonInput from "lib/components/json/JsonInput";
 import { CodeSelectSection } from "lib/components/select-code";
 import { Stepper } from "lib/components/stepper";
@@ -35,7 +38,7 @@ import {
   AmpTrackToInstantiate,
 } from "lib/services/amplitude";
 import { getCodeIdInfo } from "lib/services/code";
-import type { HumanAddr } from "lib/types";
+import type { ComposedMsg, HumanAddr } from "lib/types";
 import { AccessConfigPermission, MsgType } from "lib/types";
 import {
   composeMsg,
@@ -45,7 +48,7 @@ import {
   libDecode,
 } from "lib/utils";
 
-import { FailedModal, Footer } from "./component";
+import { Footer } from "./component";
 import type { InstantiateRedoMsg } from "./types";
 
 interface InstantiatePageState {
@@ -72,7 +75,6 @@ const Instantiate = ({ onComplete }: InstantiatePageProps) => {
   const { address = "" } = useWallet();
   const endpoint = useLCDEndpoint();
   const postInstantiateTx = useInstantiateTx();
-  const { simulate } = useSimulateFee();
   const fabricateFee = useFabricateFee();
   const { broadcast } = useTxBroadcast();
   const { validateUserAddress, validateContractAddress } = useValidateAddress();
@@ -91,7 +93,6 @@ const Instantiate = ({ onComplete }: InstantiatePageProps) => {
     formState: { errors: formErrors },
     setValue,
     watch,
-    handleSubmit,
   } = useForm<InstantiatePageState>({
     mode: "all",
     defaultValues: {
@@ -99,16 +100,9 @@ const Instantiate = ({ onComplete }: InstantiatePageProps) => {
       label: "",
       adminAddress: "",
       initMsg: "",
-      simulateError: "",
     },
   });
-
-  const {
-    codeId,
-    adminAddress: watchAdminAddress,
-    initMsg: watchInitMsg,
-    simulateError,
-  } = watch();
+  const { codeId, label, adminAddress, initMsg } = watch();
 
   const {
     control: assetsControl,
@@ -122,25 +116,38 @@ const Instantiate = ({ onComplete }: InstantiatePageProps) => {
       attachFundsOption: AttachFundsType.ATTACH_FUNDS_NULL,
     },
   });
-
   const { assetsSelect, assetsJsonStr, attachFundsOption } = watchAssets();
 
-  const disableInstantiate = useMemo(() => {
-    return (
-      !codeId ||
-      !address ||
-      !!jsonValidate(watchInitMsg) ||
-      Object.keys(formErrors).length > 0 ||
-      status.state !== "success"
-    );
-    // formErrors change doesnt trigger this effect
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [codeId, address, watchInitMsg, Object.keys(formErrors), status.state]);
+  const [composedTxMsg, setComposedTxMsg] = useState<ComposedMsg[]>([]);
+  const [estimatedFee, setEstimatedFee] = useState<StdFee>();
+  const [simulateError, setSimulateError] = useState("");
+
+  const enableInstantiate = useMemo(
+    () =>
+      !!address &&
+      codeId.length > 0 &&
+      jsonValidate(initMsg) === null &&
+      status.state === "success",
+    [address, codeId.length, initMsg, status.state]
+  );
 
   const funds = getAttachFunds({
     attachFundsOption,
     assetsJsonStr,
     assetsSelect,
+  });
+
+  const { isFetching: isSimulating } = useSimulateFeeQuery({
+    enabled: composedTxMsg.length > 0,
+    messages: composedTxMsg,
+    onSuccess: (gasRes) => {
+      if (gasRes) setEstimatedFee(fabricateFee(gasRes));
+      else setEstimatedFee(undefined);
+    },
+    onError: (e) => {
+      setSimulateError(e.message);
+      setEstimatedFee(undefined);
+    },
   });
 
   const { refetch } = useQuery(
@@ -176,51 +183,31 @@ const Instantiate = ({ onComplete }: InstantiatePageProps) => {
   // ------------------------------------------//
   // ----------------FUNCTIONS-----------------//
   // ------------------------------------------//
-  const proceed = useCallback(() => {
-    handleSubmit(async ({ adminAddress, label, initMsg }) => {
-      AmpTrackAction(AmpEvent.ACTION_EXECUTE, funds.length, attachFundsOption);
+  const proceed = useCallback(async () => {
+    AmpTrackAction(AmpEvent.ACTION_EXECUTE, funds.length, attachFundsOption);
+    const stream = await postInstantiateTx({
+      codeId: Number(codeId),
+      initMsg: JSON.parse(initMsg),
+      label,
+      admin: adminAddress,
+      funds,
+      estimatedFee,
+      onTxSucceed: onComplete,
+    });
 
-      setSimulating(true);
-
-      const msg = composeMsg(MsgType.INSTANTIATE, {
-        sender: address as HumanAddr,
-        admin: adminAddress as HumanAddr,
-        codeId: Long.fromString(codeId),
-        label,
-        msg: Buffer.from(initMsg),
-        funds,
-      });
-      try {
-        const estimatedFee = await simulate([msg]);
-        const stream = await postInstantiateTx({
-          estimatedFee: estimatedFee ? fabricateFee(estimatedFee) : undefined,
-          codeId: Number(codeId),
-          initMsg: JSON.parse(initMsg),
-          label,
-          admin: adminAddress,
-          funds,
-          onTxSucceed: onComplete,
-        });
-
-        if (stream) broadcast(stream);
-        setSimulating(false);
-      } catch (e) {
-        setValue("simulateError", (e as Error).message);
-        setSimulating(false);
-      }
-    })();
+    if (stream) broadcast(stream);
+    setSimulating(false);
   }, [
-    funds,
-    handleSubmit,
+    adminAddress,
     attachFundsOption,
-    address,
-    codeId,
-    simulate,
-    postInstantiateTx,
-    fabricateFee,
-    onComplete,
     broadcast,
-    setValue,
+    codeId,
+    estimatedFee,
+    funds,
+    initMsg,
+    label,
+    onComplete,
+    postInstantiateTx,
   ]);
 
   // ------------------------------------------//
@@ -238,6 +225,35 @@ const Instantiate = ({ onComplete }: InstantiatePageProps) => {
 
     return () => {};
   }, [address, codeId, refetch]);
+
+  useEffect(() => {
+    if (enableInstantiate) {
+      setSimulateError("");
+      const composedMsg = composeMsg(MsgType.INSTANTIATE, {
+        sender: address as HumanAddr,
+        admin: adminAddress as HumanAddr,
+        codeId: Long.fromString(codeId),
+        label,
+        msg: Buffer.from(initMsg),
+        funds,
+      });
+      const timeoutId = setTimeout(() => {
+        setComposedTxMsg([composedMsg]);
+      }, 1000);
+      return () => clearTimeout(timeoutId);
+    }
+    return () => {};
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [
+    address,
+    adminAddress,
+    codeId,
+    enableInstantiate,
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    JSON.stringify(funds),
+    initMsg,
+    label,
+  ]);
 
   useEffect(() => {
     if (codeIdQuery) setValue("codeId", codeIdQuery);
@@ -318,14 +334,14 @@ const Instantiate = ({ onComplete }: InstantiatePageProps) => {
             placeholder={`ex. ${exampleContractAddress}`}
             helperText="The contract's admin will be able to migrate and update future admins."
             variant="floating"
-            error={validateAdmin(watchAdminAddress)}
+            error={validateAdmin(adminAddress)}
             helperAction={
               <AssignMe
                 onClick={() => {
                   AmpTrack(AmpEvent.USE_ASSIGN_ME);
                   setValue("adminAddress", address);
                 }}
-                isDisable={watchAdminAddress === address}
+                isDisable={adminAddress === address}
               />
             }
           />
@@ -333,7 +349,7 @@ const Instantiate = ({ onComplete }: InstantiatePageProps) => {
             Instantiate Message
           </Heading>
           <JsonInput
-            text={watchInitMsg}
+            text={initMsg}
             setText={(newVal: string) => setValue("initMsg", newVal)}
             minLines={10}
           />
@@ -346,18 +362,38 @@ const Instantiate = ({ onComplete }: InstantiatePageProps) => {
             attachFundsOption={attachFundsOption}
           />
         </form>
+        {simulateError && (
+          <Flex gap={2} mb={4}>
+            <CustomIcon
+              name="alert-circle-solid"
+              boxSize={3}
+              color="error.main"
+            />
+            <Text variant="body3" color="error.main">
+              {simulateError}
+            </Text>
+          </Flex>
+        )}
+        <Flex
+          fontSize="14px"
+          color="text.dark"
+          alignSelf="flex-start"
+          alignItems="center"
+          display="flex"
+          gap={1}
+        >
+          <p>Transaction Fee:</p>
+          <EstimatedFeeRender
+            estimatedFee={estimatedFee}
+            loading={isSimulating}
+          />
+        </Flex>
       </WasmPageContainer>
       <Footer
         onInstantiate={proceed}
-        disabled={disableInstantiate}
+        disabled={!enableInstantiate}
         loading={simulating}
       />
-      {simulateError && (
-        <FailedModal
-          errorLog={simulateError}
-          onClose={() => setValue("simulateError", "")}
-        />
-      )}
     </>
   );
 };